autofill database clears the data for now

+++ src/main/java/com/crayons_2_0/service/database/CreateDatabase.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.crayons_2_0.authentication.CurrentUser;
 import com.crayons_2_0.model.Course;
 import com.crayons_2_0.model.CrayonsUser;
 import com.vaadin.spring.annotation.SpringComponent;
@@ -17,18 +18,23 @@ public class CreateDatabase {
     @Autowired
     private UserDAO userDAO;
     
+    CurrentUser currentUser;
+    
 	public void fillDatabase() {
 		userDAO.createTable();
 		courseDAO.createDbTable();
-		
-		List<CrayonsUser> x = userDAO.findAll();
-		for (CrayonsUser tmpUser : x) {
-			userDAO.deleteUser(tmpUser);
-		}
 		
 		List<Course> y = courseDAO.findAll();
 		for (Course tmpCrs : y) {
 			courseDAO.remove(tmpCrs);
 		}
+		
+		List<CrayonsUser> x = userDAO.findAll();
+		for (CrayonsUser tmpUser : x) {
+			if (!tmpUser.getEmail().equals(currentUser.get())){
+				userDAO.deleteUser(tmpUser);
+			}
+		}
+		
 	}
 }
